NVSHAS-7554 - Creating new files in watched directories don't trigger incidents
Looks like the issue was ultimately due to a bad regex filter.
Changed `flt.regex, _ = regexp.Compile(fmt.Sprintf("^%s$", flt.path))`
to
`flt.regex, _ = regexp.Compile(fmt.Sprintf("^%s", flt.path))`

The regex expression looks alright until we get to the `$` symbol. This
tells the regex to match "end of string". But the path we give to the
regex filter is the full linkpath and not just the directory path. So
the regex filter will always fail when creating new files in a watched
directory.

Actual definition follows
> By default, the match must occur at the end of the string or before \n
at the end of the string; in multiline mode, it must occur before the
end of the line or before \n at the end of the line.
- Source
  https://learn.microsoft.com/en-us/dotnet/standard/base-types/regular-expression-language-quick-reference