NVSHAS-7925 - Deployment stats did not make sense in UI.
I followed the paths in k8s and noticed that the cgroup file was not
did not have a valid path and fellback to /sys/fs/cgroup on Docker K8s.
When I debugged it, I noticed that the path to
/proc/<pid>/root/sys/fs/cgroup was working in this scenario. The fix I
created here will attempt to fallback to /proc/<pid>/root/sys/fs/cgroup
and check if the files exist within it. If the files do not exist, then
we will finally fallback to /sys/fs/cgroup.