NVSHAS 7990 & 7925 - cgorup error log fixes Changed getMemoryStats() to return an error for no files found.
Changed updateContainerStats() to deal with errors.
There is a check if the container is a pod holder as those errors are
valid under certain OS.
Added error handling when we first register a container in
fillContainerProperties(). If the cgroup files fail to load, we blank
out the path and never process it again. This should silence the errors
in the logs with at least one message.

Moved the cgroup path verification into GetContainerCgroupPath()
Fixed bug where the verification wasn't working on cgroup v1.
Fixed another bug where logging wasn't silenced correctly and still noisy