NVSHAS-8111 - Custom scripts not running Moved buildCustomGroups() outside of the leader check. This way if the function is called for both leader and followers. This function updates dpt.group2node set so its always up to date with the WorkloadJoin/Leave events. The transaction is moved outside of the leader check as well so we can close it if it is not a leader.
Before, if a leadership transfered happened to a follower, custom
scripts would stop getting updated because the group2node set was always
empty. We had only updated the set on the leader in the WorkloadJoin()
function. So followers wouldn't update their set until all the clients
rejoined the cluster.

WorkloadLeave() didn't have this issue because it called
dpt.purgeCustomGroupsByNodeGroups() which would clean group2node for
both leader & follower.

refreshServiceGroup2Nodes() looks like it should update the group2nodes
set but custom groups are part of the function arguments so it can't
refresh it implicitly. But this would compete with the leader calling
dpt.buildCustomGroups(). Not the end of the world but it would be
redundant on the leader maybe.