NVSHAS-7947 - Process violation is reported as Host.Suspicious instead Container.Suspicious
I believe this is a race condition where the container map being
updated at the same time we are evaluating a new process. The map data
is not available so it assumes it is a host process.

I made a change in evaluateApplication() so when it detects a violation,
it will make a last ditch attempt to confirm the process is not a host
process by checking:
1. if pidContainerMap has the pid and has an updated information about
   the container
2. Make another attempt at GetContainerIDByPID() to determine the
   container ID from the PID.

I also added some extra information to recordIncident() because the log
entry was not useful in debugging as is. Either we make it more useful
or we remove it.