NVSHAS-8502 properly generate installation id
In the previous version, installation ID can fail to be created and
cause trouble.

The reason is that after a fresh install, controllers will compete for
leader.  All write operations in consul is strong consistent, that
means once leader is changed, any write operation will fail with
"leadership lost while committing log" error.

This commit fixes the issue by making the logic a lazy initialization,
so even when consul is temporarily offline, the later attempts will
succeed and reset installation ID when necessary.