NVSHAS-8613 repeated sigsegv in one controller
It's likely that the oldest/leader controller hit a race condition,
which resulted in a corrupted audit event cache.

When a controller starts, as its initialization logic, it will sync with
other controllers.  The container can crash when it reads the corrupted
audit event cache via syncAuditRx() call.

The reason why there is only one controller crashing is probably because
the other two controllers didn't call the cache.syncAuditRx().

Among those two containers that didn't crash, one container is the
leader container.  As long as the other has finished the
learning/resync, the issue will not happen.

It's not 100% certain why the controller crashed in the first place, but
that is something that can be recovered by monitor for sure.

In this commit, a proper protection is added before calling
benchUpdate() and auditUpdate().

Another check is also added in Rx side, so already corrupted data can be
corrected.