fix: NVSHAS-8968 resp.Body is not closed
When docker hub's token service is called too many times, backend would
start throttling and return 429 too many requests.

When this happens, the response from the token service will not be
closed and cause memory leak gradually.

While writing unit-test, it was found that the second request's error
will always be nil due to the err variable is shadowed.

This issue is also fixed and covered in unit tests.