fix: NVSHAS-9342 controller failed to load secret
This happens when a secret is being updated and a NV component is
starting. In this case, NV component could read an empty internal
secret, which is as expected.  However, the current logic doesn't
ignore this secret correctly, hence the issue.

This commit makes sure that informer will wait until certificate is
available in the secret.